deps: update V8 to 10.9.194.4
PR-URL: https://github.com/nodejs/node/pull/45579

+++ deps/v8/src/codegen/loong64/constants-loong64.h
@@ -693,7 +693,8 @@ inline Hint NegateHint(Hint hint) { return no_hint; }
 // registers and other constants.
 
 // Break 0xfffff, reserved for redirected real time call.
-const Instr rtCallRedirInstr = BREAK | call_rt_redirected;
+const Instr rtCallRedirInstr =
+    static_cast<uint32_t>(BREAK) | call_rt_redirected;
 // A nop instruction. (Encoding of addi_w 0 0 0).
 const Instr nopInstr = ADDI_W;
